requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.3
deepl==1.16.1
distro==1.8.0
exceptiongroup==1.2.0
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
keyboard==0.13.5
langchain==0.0.345
langchain-core==0.0.9
langsmith==0.0.69
marshmallow==3.20.1
MouseInfo==0.1.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.3.7
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.54
pydantic==2.5.2
pydantic_core==2.14.5
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyocr==0.8.5
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.0
pytweening==1.0.7
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.3
